- Fixed some warnings using XCode 5.0 on MacOSX 10.8.

+++ include/cybergarage/soap/SOAPRequest.h
@@ -95,7 +95,7 @@ public:
 
 private:
 
-	CyberXML::Node *parseMessage(const std::string &content, int contentLen);
+	CyberXML::Node *parseMessage(const std::string &content, size_t contentLen);
 
 public:
